Feat :: 회원가입 로직을 AuthRequestManager에서 처리
