+++ app/src/main/java/com/example/palette/ui/register/JoinCompleteFragment.kt
@@ -4,30 +4,16 @@ import android.graphics.Color
 import android.graphics.LinearGradient
 import android.graphics.Shader
 import android.os.Bundle
-import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.fragment.app.activityViewModels
-import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.findNavController
 import com.example.palette.R
-import com.example.palette.application.PaletteApplication
-import com.example.palette.common.Constant.TAG
-import com.example.palette.common.HeaderUtil
-import com.example.palette.data.auth.RegisterRequest
-import com.example.palette.data.auth.RegisterRequestManager
 import com.example.palette.databinding.FragmentJoinCompleteBinding
-import com.example.palette.ui.util.shortToast
-import kotlinx.coroutines.SupervisorJob
-import kotlinx.coroutines.launch
-import retrofit2.HttpException
-import java.net.SocketTimeoutException
 
 class JoinCompleteFragment : Fragment() {
     private lateinit var binding : FragmentJoinCompleteBinding
-    private val registerViewModel: RegisterViewModel by activityViewModels()
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
